Feature request (test)
I sent a mail to the developer suggesting that a right clicking (long clicking on Android) on a field that already contains a number would do the same as left/short clicking instead of doing nothing. That would reduce the number of times the short/long click toggle had to be clicked since long clicking on small squares is difficult, but short clicking on numbers and long clicking on empty squares are the most common actions (at least for me).
I'm just guessing this change will do what I want since I didn't program this myself. But it would be cool if it actually worked.